requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.4
django-ratelimit==4.1.0
django_celery_results==2.6.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.11
kombu==5.6.2
packaging==26.1
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
